backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
newsapi==0.1.1
requests==2.32.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
yfinance==0.2.28
